requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpickle==1.1
MarkupSafe==1.1.1
pycryptodome==3.8.1
tabulate==0.8.3
Werkzeug==0.15.2
